Fixed layouts and reworked UserDropDown
